feat: `ant` logs to stderr
BREAKING CHANGE: the `--log-output-dest` argument is removed. Logging to file is a bit strange for a
short-lived program, so we now log to stderr. The user can redirect to file if they want to.

Logging is off by default. It is switched on by using the `verbosity` argument in the standard `-v`
to `-vvvvv` form. This increases the levels from `WARN` to `TRACE`.

The default logging targets were changed to crates that seem a bit more relevant for using `ant`,
and to remove particularly noisy crates, like `ant_bootstrap`. All the crates will log at the same
level, as defined by the `verbosity` argument; previously the default logging targets used different
levels for different crates, and those seemed a little arbitrary.

The behaviour of `ANT_LOG` is retained, so the user can still get output for particular crates, or
use the special `all` option to get trace-level logging for all crates.