chore: remove the `verbosity` argument from `antctl`
This was an argument that was implemented in a clumsy manual way that was intended to vary the
amount of output on the node manager. We will need to provide another `verbose` argument on the
`local run` command to control the output of the logs for local networks.

We still need to retain the use of it, but instead of setting it via a command line argument, it
will just be set to `Normal`. The `node-launchpad` sets it to `Minimal` in the places where it needs
the text output reduced.