feat: provide `verbosity` arg for `local run` cmd
This also provides the `-v` mechanism for varying the logging output for the local network. The
value that is set here is passed to each of the node processes.

The logging output destination is to file, otherwise on the terminal you would get a blizzard of
output from all the node processes.